# specs/007-store-access-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: BurgoOS Store Access Management API
  version: 0.1.0
  description: Tenant-scoped admin contracts for login, users, access profiles, permissions and access audit.
paths:
...
  /api/admin/access/users:
    get:
      summary: List users visible to the current actor
      parameters:
        - name: storeId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/UserStatus"
        - name: profileId
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Scoped user list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
    post:
      summary: Create a user within the actor's allowed scope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
        "403":
          description: Actor cannot assign requested store or profile
        "409":
          description: Login already exists
  /api/admin/access/users/{id}:
    get:
      summary: Get scoped user detail
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "403":
          description: User is outside actor scope
    patch:
      summary: Update a user, status or assignments within actor scope
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
        "403":
          description: Actor cannot perform requested change
        "409":
          description: Change would violate master or uniqueness rules
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    UserStatus:
      type: string
      enum: [INVITED, ACTIVE, INACTIVE, LOCKED]
    ProfileStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    ProfileScope:
      type: string
      enum: [GLOBAL, STORE]
...
    StoreSummary:
      type: object
      required: [id, name, slug, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
    UserSummary:
      type: object
      required: [id, login, name, status, isMaster]
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        status:
          $ref: "#/components/schemas/UserStatus"
        isMaster:
          type: boolean
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/UserSummary"
        - type: object
          required: [assignments]
          properties:
            phone:
              type: string
              nullable: true
            lastLoginAt:
              type: string
              format: date-time
              nullable: true
            assignments:
              type: array
              items:
                $ref: "#/components/schemas/UserStoreAssignment"
    UserInput:
      type: object
      required: [login, name, email, isMaster, assignments]
      properties:
        login:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        isMaster:
          type: boolean
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/UserStoreAssignmentInput"
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/UserStatus"
        isMaster:
          type: boolean
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/UserStoreAssignmentInput"
    UserStoreAssignment:
      type: object
      required: [store, profile, canManageStoreAccess, status]
      properties:
        store:
          $ref: "#/components/schemas/StoreSummary"
        profile:
          $ref: "#/components/schemas/AccessProfileSummary"
        canManageStoreAccess:
          type: boolean
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
    UserStoreAssignmentInput:
      type: object
      required: [storeId, profileId, canManageStoreAccess, status]
      properties:
        storeId:
          type: string
          format: uuid
        profileId:
          type: string
          format: uuid
        canManageStoreAccess:
          type: boolean
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
    AccessProfileSummary:
      type: object
      required: [id, name, scope, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        scope:
          $ref: "#/components/schemas/ProfileScope"
        storeId:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: "#/components/schemas/ProfileStatus"
...